DANGEROUS - read Desc
Just have a previous pointer for changing the links when elements in node are repeated

And dummy node is used just for the case when the start itself is repeat as in that case we cant have a previous since it doesnt exists in the linked list.
Eg. the case 1->1->1->2->3

Also return dummy->next since it has to return the next value pointed
(Case when whole list becomes empty Eg. 1->1->1->2->2)